move to pure header implementation
This is required for the templates to work correctly.
